requirements.txt
Django==3.2.9
pydash==5.1.0
djangorestframework==3.12.4
djangorestframework_simplejwt==5.0.0
PyJWT==2.3.0
psycopg2==2.9.2
cryptography==36.0.0
